+++ example/sm_pairing_peripheral.c
@@ -93,12 +93,14 @@ static void sm_peripheral_setup(void){
 
     /**
      * Choose ONE of the following configurations
-     * Bonding is disabled to allow for repeated testing. It can be enabled with SM_AUTHREQ_BONDING 
+     * Bonding is disabled to allow for repeated testing. It can be enabled by or'ing
+     * SM_AUTHREQ_BONDING to the authentication requirements like this:
+     * sm_set_authentication_requirements( X | SM_AUTHREQ_BONDING)
      */
 
     // LE Legacy Pairing, Just Works
     // sm_set_io_capabilities(IO_CAPABILITY_NO_INPUT_NO_OUTPUT);
-    // sm_set_authentication_requirements(SM_AUTHREQ_BONDING);
+    // sm_set_authentication_requirements(0);
 
     // LE Legacy Pairing, Passkey entry initiator enter, responder (us) displays
     // sm_set_io_capabilities(IO_CAPABILITY_DISPLAY_ONLY);
@@ -106,12 +108,16 @@ static void sm_peripheral_setup(void){
     // sm_use_fixed_passkey_in_display_role(123456);
 
 #ifdef ENABLE_LE_SECURE_CONNECTIONS
-    // LE Secure Connetions, Just Works
+
+    // enable LE Secure Connections Only mode - disables Legacy pairing
+    sm_set_secure_connections_only_mode(true);
+
+    // LE Secure Connections, Just Works
     // sm_set_io_capabilities(IO_CAPABILITY_DISPLAY_YES_NO);
     // sm_set_authentication_requirements(SM_AUTHREQ_SECURE_CONNECTION);
 
     // LE Secure Connections, Numeric Comparison
-    sm_set_io_capabilities(IO_CAPABILITY_DISPLAY_ONLY);
+    sm_set_io_capabilities(IO_CAPABILITY_DISPLAY_YES_NO);
     sm_set_authentication_requirements(SM_AUTHREQ_SECURE_CONNECTION|SM_AUTHREQ_MITM_PROTECTION);
 
     // LE Legacy Pairing, Passkey entry initiator enter, responder (us) displays
@@ -207,7 +213,7 @@ static void packet_handler (uint8_t packet_type, uint16_t channel, uint8_t *pack
                             printf("Pairing failed, timeout\n");
                             break;
                         case ERROR_CODE_REMOTE_USER_TERMINATED_CONNECTION:
-                            printf("Pairing faileed, disconnected\n");
+                            printf("Pairing failed, disconnected\n");
                             break;
                         case ERROR_CODE_AUTHENTICATION_FAILURE:
                             printf("Pairing failed, reason = %u\n", sm_event_pairing_complete_get_reason(packet));